Segmentation_OPV2V/requirements.txt
addict==2.4.0
asttokens==3.0.0
attrs==25.3.0
autocommand==2.2.2
backcall==0.2.0
backports.tarfile==1.2.0
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
ccimport==0.4.4
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cumm==0.5.3
cycler==0.12.1
Cython==3.0.12
dash==3.0.0
decorator==5.2.1
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
fonttools==4.56.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.3.1
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
more-itertools==10.3.0
narwhals==1.31.0
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.11.1.4
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
open3d==0.19.0
-e git+ssh://git@github.com/shilpa2301/CoBEVT_AutoNet.git@01b19f321200527456ce51b3c64f91a4749fa5e7#egg=OpenCOOD&subdirectory=opv2v
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
parso==0.8.4
pccm==0.4.16
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
portalocker==3.0.0
prompt_toolkit==3.0.50
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.1
pyparsing==3.1.4
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests @ file:///croot/requests_1721410876868/work
retrying==1.3.4
rpds-py==0.20.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
Shapely==1.8.5
six==1.17.0
stack-data==0.6.3
stringcase==1.2.0
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.15
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.20.2
cython
